Move to string subpackage
