+++ johnzon-jsonb/src/test/java/org/apache/johnzon/jsonb/symmetry/adapter/string/StringAdapterPrecedenceConfigClassGetterFieldConstructorHasSetterTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.johnzon.jsonb.symmetry.adapter;
+package org.apache.johnzon.jsonb.symmetry.adapter.string;
 
 import jakarta.json.bind.Jsonb;
 import jakarta.json.bind.JsonbBuilder;
@@ -27,19 +27,21 @@ import static org.junit.Assert.assertEquals;
 
 /**
  * JsonbTypeAdapter on
- *  - Field
- *  - Constructor
- *  - Getter
- *  - Setter
  *  - Config
  *  - Class
+ *  - Constructor
+ *  - Getter
+ *  - Field
  *
- *  Setter wins on read
- *  Getter wins on write
+ * Still has a setter
  *
- *  Constructor adapter is called, but overwritten
+ * Outcome
+ *
+ *  - Field wins on read
+ *  - Getter wins on write
+ *  - Constructor adapter is called, but overwritten
  */
-public class StringAdapterPrecedenceConfigClassGetterSetterFieldConstructorTest extends StringAdapterOnClassTest {
+public class StringAdapterPrecedenceConfigClassGetterFieldConstructorHasSetterTest extends StringAdapterOnClassTest {
 
     @Override
     public Jsonb jsonb() {
@@ -50,10 +52,10 @@ public class StringAdapterPrecedenceConfigClassGetterSetterFieldConstructorTest 
     public void assertRead(final Jsonb jsonb) {
         final String json = "{\"email\":\"test@domain.com\"}";
         final Contact actual = jsonb.fromJson(json, Contact.class);
-        assertEquals("Contact{email=test@domain.com:Setter.adaptFromJson}", actual.toString());
+        assertEquals("Contact{email=test@domain.com:Field.adaptFromJson}", actual.toString());
         assertEquals("Constructor.adaptFromJson\n" +
                 "Contact.<init>\n" +
-                "Setter.adaptFromJson\n" +
+                "Field.adaptFromJson\n" +
                 "Contact.setEmail", calls());
     }
 
@@ -86,7 +88,6 @@ public class StringAdapterPrecedenceConfigClassGetterSetterFieldConstructorTest 
             return email;
         }
 
-        @JsonbTypeAdapter(Adapter.Setter.class)
         public void setEmail(final Email email) {
             CALLS.called();
             this.email = email;
